pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-quest"
version = "0.1.0"
description = "AI-powered D&D-style text RPG with LightRAG knowledge graph backend"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "RAG-Quest Contributors"},
]
keywords = ["rpg", "game", "ai", "rag", "knowledge-graph", "text-adventure"]
license = {text = "MIT"}

dependencies = [
    "lightrag-hku>=1.0",
    "httpx>=0.27",
    "rich>=13.0",
    "pymupdf>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
]

[project.scripts]
rag-quest = "rag_quest.__main__:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
